\name{rLSparam}
\alias{rLSparam}

\alias{LS_MIN}
\alias{LS_MAX}
\alias{LS_CONTYPE_GE}
\alias{LS_CONTYPE_LE}
\alias{LS_CONTYPE_EQ}
\alias{LS_CONTYPE_FR}
\alias{LS_CONETYPE_QUAD}
\alias{LS_CONETYPE_RQUAD}
\alias{LS_VARTYPE_CONT}
\alias{LS_VARTYPE_BIN}
\alias{LS_VARTYPE_INT}
\alias{LS_INFINITY}
\alias{LS_BASTYPE_BAS}
\alias{LS_BASTYPE_ATLO}
\alias{LS_BASTYPE_ATUP}
\alias{LS_BASTYPE_FNUL}
\alias{LS_BASTYPE_SBAS}
\alias{LS_UNFORMATTED_MPS}
\alias{LS_FORMATTED_MPS}
\alias{LS_UNFORMATTED_MPS_COMP}
\alias{LS_FORMATTED_MPS_COMP}
\alias{LS_SOLUTION_OPT}
\alias{LS_SOLUTION_MIP}
\alias{LS_SOLUTION_OPT_IPM}
\alias{LS_SOLUTION_OPT_OLD}
\alias{LS_SOLUTION_MIP_OLD}
\alias{LS_BASFILE_BIN}
\alias{LS_BASFILE_MPS}
\alias{LS_BASFILE_TXT}
\alias{LS_INT_TYPE}
\alias{LS_DOUBLE_TYPE}
\alias{LS_MAX_ERROR_MESSAGE_LENGTH}
\alias{LS_DEFAULT}
\alias{LS_MAX_JOBJECTS}
\alias{LS_PROPERTY_UNKNOWN}
\alias{LS_PROPERTY_CONST}
\alias{LS_PROPERTY_LINEAR}
\alias{LS_PROPERTY_CONVEX}
\alias{LS_PROPERTY_CONCAVE}
\alias{LS_PROPERTY_QUASI_CONVEX}
\alias{LS_PROPERTY_QUASI_CONCAVE}
\alias{LS_PROPERTY_MAX}
\alias{LS_PROPERTY_MONO_INCREASE}
\alias{LS_PROPERTY_MONO_DECREASE}
\alias{LS_RAW_COPY}
\alias{LS_DEEP_COPY}
\alias{LS_TIME_COPY}
\alias{LS_STOC_COPY}
\alias{LS_SNGSTG_COPY}
\alias{LSSEC01}
\alias{LSSEC02}
\alias{LSSEC03}
\alias{LSSEC04}
\alias{LSSEC05}
\alias{LSSEC06}
\alias{LSSEC10}
\alias{LSSEC15}
\alias{LSSEC20}
\alias{LSSEC30}
\alias{LSMIN01}
\alias{LSMIN02}
\alias{LSMIN03}
\alias{LSMIN05}
\alias{LSMIN06}
\alias{LSMIN10}
\alias{LSMIN15}
\alias{LSMIN20}
\alias{LSMIN30}
\alias{LSHOUR01}
\alias{LSHOUR02}
\alias{LSHOUR03}
\alias{LSHOUR05}
\alias{LSHOUR06}
\alias{LSHOUR08}
\alias{LSHOUR12}
\alias{LSDAY}
\alias{LSWEEK}
\alias{LSMONTH}
\alias{LSQUARTER}
\alias{LSYEAR}
\alias{LSSUNDAY}
\alias{LSMONDAY}
\alias{LSTUESDAY}
\alias{LSWEDNESDAY}
\alias{LSTHURSDAY}
\alias{LSFRIDAY}
\alias{LSSATURDAY}
\alias{LS_DATA_CORE}
\alias{LS_DATA_TIME}
\alias{LS_DATA_STOC}
\alias{LS_DATA_FILE}
\alias{LS_STATUS_OPTIMAL}
\alias{LS_STATUS_BASIC_OPTIMAL}
\alias{LS_STATUS_INFEASIBLE}
\alias{LS_STATUS_UNBOUNDED}
\alias{LS_STATUS_FEASIBLE}
\alias{LS_STATUS_INFORUNB}
\alias{LS_STATUS_NEAR_OPTIMAL}
\alias{LS_STATUS_LOCAL_OPTIMAL}
\alias{LS_STATUS_LOCAL_INFEASIBLE}
\alias{LS_STATUS_CUTOFF}
\alias{LS_STATUS_NUMERICAL_ERROR}
\alias{LS_STATUS_UNKNOWN}
\alias{LS_STATUS_UNLOADED}
\alias{LS_STATUS_LOADED}
\alias{LS_STATUS_BOUNDED}
\alias{LS_IPARAM_OBJSENSE}
\alias{LS_DPARAM_CALLBACKFREQ}
\alias{LS_DPARAM_OBJPRINTMUL}
\alias{LS_IPARAM_CHECK_FOR_ERRORS}
\alias{LS_IPARAM_ALLOW_CNTRLBREAK}
\alias{LS_IPARAM_DECOMPOSITION_TYPE}
\alias{LS_IPARAM_LP_SCALE}
\alias{LS_IPARAM_LP_ITRLMT}
\alias{LS_IPARAM_SPLEX_PPRICING}
\alias{LS_IPARAM_SPLEX_REFACFRQ}
\alias{LS_IPARAM_BARRIER_SOLVER}
\alias{LS_IPARAM_PROB_TO_SOLVE}
\alias{LS_IPARAM_LP_PRINTLEVEL}
\alias{LS_IPARAM_MPS_OBJ_WRITESTYLE}
\alias{LS_IPARAM_SPLEX_DPRICING}
\alias{LS_IPARAM_SOL_REPORT_STYLE}
\alias{LS_IPARAM_INSTRUCT_LOADTYPE}
\alias{LS_IPARAM_SPLEX_DUAL_PHASE}
\alias{LS_IPARAM_LP_PRELEVEL}
\alias{LS_IPARAM_STRING_LENLMT}
\alias{LS_IPARAM_USE_NAMEDATA}
\alias{LS_IPARAM_SPLEX_USE_EXTERNAL}
\alias{LS_DPARAM_LP_ITRLMT}
\alias{LS_IPARAM_COPY_MODE}
\alias{LS_IPARAM_MULTITHREAD_MODE}
\alias{LS_IPARAM_FIND_BLOCK}
\alias{LS_IPARAM_PROFILER_LEVEL}
\alias{LS_IPARAM_INSTRUCT_READMODE}
\alias{LS_IPARAM_INSTRUCT_SUBOUT}
\alias{LS_IPARAM_SOLVER_IUSOL}
\alias{LS_IPARAM_SOLVER_TIMLMT}
\alias{LS_DPARAM_SOLVER_CUTOFFVAL}
\alias{LS_DPARAM_SOLVER_FEASTOL}
\alias{LS_IPARAM_SOLVER_RESTART}
\alias{LS_IPARAM_SOLVER_IPMSOL}
\alias{LS_DPARAM_SOLVER_OPTTOL}
\alias{LS_IPARAM_SOLVER_USECUTOFFVAL}
\alias{LS_IPARAM_SOLVER_PRE_ELIM_FILL}
\alias{LS_DPARAM_SOLVER_TIMLMT}
\alias{LS_IPARAM_SOLVER_CONCURRENT_OPTMODE}
\alias{LS_DPARAM_SOLVER_PERT_FEASTOL}
\alias{LS_IPARAM_SOLVER_PARTIALSOL_LEVEL}
\alias{LS_DPARAM_LP_MIN_FEASTOL}
\alias{LS_DPARAM_LP_MAX_FEASTOL}
\alias{LS_DPARAM_LP_MIN_OPTTOL}
\alias{LS_DPARAM_LP_MAX_OPTTOL}
\alias{LS_DPARAM_LP_MIN_PIVTOL}
\alias{LS_DPARAM_LP_MAX_PIVTOL}
\alias{LS_DPARAM_LP_AIJ_ZEROTOL}
\alias{LS_DPARAM_LP_PIV_ZEROTOL}
\alias{LS_DPARAM_LP_PIV_BIGTOL}
\alias{LS_DPARAM_LP_BIGM}
\alias{LS_DPARAM_LP_BNDINF}
\alias{LS_DPARAM_LP_INFINITY}
\alias{LS_IPARAM_LP_PPARTIAL}
\alias{LS_IPARAM_LP_DPARTIAL}
\alias{LS_IPARAM_LP_DRATIO}
\alias{LS_IPARAM_LP_PRATIO}
\alias{LS_IPARAM_LP_RATRANGE}
\alias{LS_IPARAM_LP_DPSWITCH}
\alias{LS_IPARAM_LP_PALLOC}
\alias{LS_IPARAM_LP_PRTFG}
\alias{LS_IPARAM_LP_OPRFREE}
\alias{LS_IPARAM_LP_SPRINT_SUB}
\alias{LS_IPARAM_LU_NUM_CANDITS}
\alias{LS_IPARAM_LU_MAX_UPDATES}
\alias{LS_IPARAM_LU_PRINT_LEVEL}
\alias{LS_IPARAM_LU_UPDATE_TYPE}
\alias{LS_IPARAM_LU_PIVMOD}
\alias{LS_DPARAM_LU_EPS_DIAG}
\alias{LS_DPARAM_LU_EPS_NONZ}
\alias{LS_DPARAM_LU_EPS_PIVABS}
\alias{LS_DPARAM_LU_EPS_PIVREL}
\alias{LS_DPARAM_LU_INI_RCOND}
\alias{LS_DPARAM_LU_SPVTOL_UPDATE}
\alias{LS_DPARAM_LU_SPVTOL_FTRAN}
\alias{LS_DPARAM_LU_SPVTOL_BTRAN}
\alias{LS_DPARAM_IPM_TOL_INFEAS}
\alias{LS_DPARAM_IPM_TOL_PATH}
\alias{LS_DPARAM_IPM_TOL_PFEAS}
\alias{LS_DPARAM_IPM_TOL_REL_STEP}
\alias{LS_DPARAM_IPM_TOL_PSAFE}
\alias{LS_DPARAM_IPM_TOL_DFEAS}
\alias{LS_DPARAM_IPM_TOL_DSAFE}
\alias{LS_DPARAM_IPM_TOL_MU_RED}
\alias{LS_DPARAM_IPM_BASIS_REL_TOL_S}
\alias{LS_DPARAM_IPM_BASIS_TOL_S}
\alias{LS_DPARAM_IPM_BASIS_TOL_X}
\alias{LS_DPARAM_IPM_BI_LU_TOL_REL_PIV}
\alias{LS_DPARAM_IPM_CO_TOL_INFEAS}
\alias{LS_DPARAM_IPM_CO_TOL_PFEAS}
\alias{LS_DPARAM_IPM_CO_TOL_DFEAS}
\alias{LS_DPARAM_IPM_CO_TOL_MU_RED}
\alias{LS_IPARAM_IPM_MAX_ITERATIONS}
\alias{LS_IPARAM_IPM_OFF_COL_TRH}
\alias{LS_IPARAM_IPM_CHECK_CONVEXITY}
\alias{LS_IPARAM_NLP_SOLVE_AS_LP}
\alias{LS_IPARAM_NLP_SOLVER}
\alias{LS_IPARAM_NLP_SUBSOLVER}
\alias{LS_IPARAM_NLP_PRINTLEVEL}
\alias{LS_DPARAM_NLP_PSTEP_FINITEDIFF}
\alias{LS_IPARAM_NLP_DERIV_DIFFTYPE}
\alias{LS_DPARAM_NLP_FEASTOL}
\alias{LS_DPARAM_NLP_REDGTOL}
\alias{LS_IPARAM_NLP_USE_CRASH}
\alias{LS_IPARAM_NLP_USE_STEEPEDGE}
\alias{LS_IPARAM_NLP_USE_SLP}
\alias{LS_IPARAM_NLP_USE_SELCONEVAL}
\alias{LS_IPARAM_NLP_PRELEVEL}
\alias{LS_IPARAM_NLP_ITRLMT}
\alias{LS_IPARAM_NLP_LINEARZ}
\alias{LS_IPARAM_NLP_LINEARITY}
\alias{LS_IPARAM_NLP_STARTPOINT}
\alias{LS_IPARAM_NLP_CONVEXRELAX}
\alias{LS_IPARAM_NLP_CR_ALG_REFORM}
\alias{LS_IPARAM_NLP_QUADCHK}
\alias{LS_IPARAM_NLP_AUTODERIV}
\alias{LS_IPARAM_NLP_MAXLOCALSEARCH}
\alias{LS_IPARAM_NLP_CONVEX}
\alias{LS_IPARAM_NLP_CONOPT_VER}
\alias{LS_IPARAM_NLP_USE_LINDO_CRASH}
\alias{LS_IPARAM_NLP_STALL_ITRLMT}
\alias{LS_IPARAM_NLP_AUTOHESS}
\alias{LS_IPARAM_NLP_FEASCHK}
\alias{LS_DPARAM_NLP_ITRLMT}
\alias{LS_IPARAM_NLP_MAXSUP}
\alias{LS_IPARAM_NLP_MSW_SOLIDX}
\alias{LS_IPARAM_NLP_ITERS_PER_LOGLINE}
\alias{LS_IPARAM_NLP_MAX_RETRY}
\alias{LS_IPARAM_NLP_MSW_NORM}
\alias{LS_IPARAM_NLP_MSW_POPSIZE}
\alias{LS_IPARAM_NLP_MSW_MAXPOP}
\alias{LS_IPARAM_NLP_MSW_MAXNOIMP}
\alias{LS_IPARAM_NLP_MSW_FILTMODE}
\alias{LS_DPARAM_NLP_MSW_POXDIST_THRES}
\alias{LS_DPARAM_NLP_MSW_EUCDIST_THRES}
\alias{LS_DPARAM_NLP_MSW_XNULRAD_FACTOR}
\alias{LS_DPARAM_NLP_MSW_XKKTRAD_FACTOR}
\alias{LS_IPARAM_NLP_MAXLOCALSEARCH_TREE}
\alias{LS_IPARAM_NLP_MSW_RG_SEED}
\alias{LS_IPARAM_NLP_MSW_PREPMODE}
\alias{LS_IPARAM_NLP_MSW_RMAPMODE}
\alias{LS_IPARAM_NLP_XSMODE}
\alias{LS_DPARAM_NLP_MSW_OVERLAP_RATIO}
\alias{LS_DPARAM_NLP_INF}
\alias{LS_IPARAM_NLP_IPM2GRG}
\alias{LS_IPARAM_NLP_USE_SDP}
\alias{LS_IPARAM_MIP_TIMLIM}
\alias{LS_IPARAM_MIP_AOPTTIMLIM}
\alias{LS_IPARAM_MIP_LSOLTIMLIM}
\alias{LS_IPARAM_MIP_PRELEVEL}
\alias{LS_IPARAM_MIP_NODESELRULE}
\alias{LS_DPARAM_MIP_INTTOL}
\alias{LS_DPARAM_MIP_RELINTTOL}
\alias{LS_DPARAM_MIP_RELOPTTOL}
\alias{LS_DPARAM_MIP_PEROPTTOL}
\alias{LS_IPARAM_MIP_MAXCUTPASS_TOP}
\alias{LS_IPARAM_MIP_MAXCUTPASS_TREE}
\alias{LS_DPARAM_MIP_ADDCUTPER}
\alias{LS_DPARAM_MIP_ADDCUTPER_TREE}
\alias{LS_IPARAM_MIP_MAXNONIMP_CUTPASS}
\alias{LS_IPARAM_MIP_CUTLEVEL_TOP}
\alias{LS_IPARAM_MIP_CUTLEVEL_TREE}
\alias{LS_IPARAM_MIP_CUTTIMLIM}
\alias{LS_IPARAM_MIP_CUTDEPTH}
\alias{LS_IPARAM_MIP_CUTFREQ}
\alias{LS_IPARAM_MIP_HEULEVEL}
\alias{LS_IPARAM_MIP_PRINTLEVEL}
\alias{LS_IPARAM_MIP_PREPRINTLEVEL}
\alias{LS_DPARAM_MIP_CUTOFFOBJ}
\alias{LS_IPARAM_MIP_USECUTOFFOBJ}
\alias{LS_IPARAM_MIP_STRONGBRANCHLEVEL}
\alias{LS_IPARAM_MIP_TREEREORDERLEVEL}
\alias{LS_IPARAM_MIP_BRANCHDIR}
\alias{LS_IPARAM_MIP_TOPOPT}
\alias{LS_IPARAM_MIP_REOPT}
\alias{LS_IPARAM_MIP_SOLVERTYPE}
\alias{LS_IPARAM_MIP_KEEPINMEM}
\alias{LS_IPARAM_MIP_BRANCHRULE}
\alias{LS_DPARAM_MIP_REDCOSTFIX_CUTOFF}
\alias{LS_DPARAM_MIP_ADDCUTOBJTOL}
\alias{LS_IPARAM_MIP_HEUMINTIMLIM}
\alias{LS_IPARAM_MIP_BRANCH_PRIO}
\alias{LS_IPARAM_MIP_SCALING_BOUND}
\alias{LS_DPARAM_MIP_PSEUDOCOST_WEIGT}
\alias{LS_DPARAM_MIP_LBIGM}
\alias{LS_DPARAM_MIP_DELTA}
\alias{LS_IPARAM_MIP_DUAL_SOLUTION}
\alias{LS_IPARAM_MIP_BRANCH_LIMIT}
\alias{LS_DPARAM_MIP_ITRLIM}
\alias{LS_IPARAM_MIP_AGGCUTLIM_TOP}
\alias{LS_IPARAM_MIP_AGGCUTLIM_TREE}
\alias{LS_DPARAM_MIP_SWITCHFAC_SIM_IPM_ITER}
\alias{LS_IPARAM_MIP_ANODES_SWITCH_DF}
\alias{LS_DPARAM_MIP_ABSOPTTOL}
\alias{LS_DPARAM_MIP_MINABSOBJSTEP}
\alias{LS_IPARAM_MIP_PSEUDOCOST_RULE}
\alias{LS_IPARAM_MIP_ENUM_HEUMODE}
\alias{LS_IPARAM_MIP_PRELEVEL_TREE}
\alias{LS_DPARAM_MIP_REDCOSTFIX_CUTOFF_TREE}
\alias{LS_IPARAM_MIP_USE_INT_ZERO_TOL}
\alias{LS_IPARAM_MIP_USE_CUTS_HEU}
\alias{LS_DPARAM_MIP_BIGM_FOR_INTTOL}
\alias{LS_IPARAM_MIP_STRONGBRANCHDONUM}
\alias{LS_IPARAM_MIP_MAKECUT_INACTIVE_COUNT}
\alias{LS_IPARAM_MIP_PRE_ELIM_FILL}
\alias{LS_IPARAM_MIP_HEU_MODE}
\alias{LS_DPARAM_MIP_TIMLIM}
\alias{LS_DPARAM_MIP_AOPTTIMLIM}
\alias{LS_DPARAM_MIP_LSOLTIMLIM}
\alias{LS_DPARAM_MIP_CUTTIMLIM}
\alias{LS_DPARAM_MIP_HEUMINTIMLIM}
\alias{LS_IPARAM_MIP_FP_MODE}
\alias{LS_DPARAM_MIP_FP_WEIGHT}
\alias{LS_IPARAM_MIP_FP_OPT_METHOD}
\alias{LS_DPARAM_MIP_FP_TIMLIM}
\alias{LS_IPARAM_MIP_FP_ITRLIM}
\alias{LS_IPARAM_MIP_FP_HEU_MODE}
\alias{LS_DPARAM_MIP_OBJ_THRESHOLD}
\alias{LS_IPARAM_MIP_LOCALBRANCHNUM}
\alias{LS_DPARAM_MIP_SWITCHFAC_SIM_IPM_TIME}
\alias{LS_DPARAM_MIP_ITRLIM_SIM}
\alias{LS_DPARAM_MIP_ITRLIM_NLP}
\alias{LS_DPARAM_MIP_ITRLIM_IPM}
\alias{LS_IPARAM_MIP_MAXNUM_MIP_SOL_STORAGE}
\alias{LS_IPARAM_MIP_CONCURRENT_TOPOPTMODE}
\alias{LS_IPARAM_MIP_CONCURRENT_REOPTMODE}
\alias{LS_IPARAM_MIP_PREHEU_LEVEL}
\alias{LS_IPARAM_MIP_PREHEU_PRE_LEVEL}
\alias{LS_IPARAM_MIP_PREHEU_PRINT_LEVEL}
\alias{LS_IPARAM_MIP_PREHEU_TC_ITERLIM}
\alias{LS_IPARAM_MIP_PREHEU_DFE_VSTLIM}
\alias{LS_IPARAM_MIP_PREHEU_VAR_SEQ}
\alias{LS_IPARAM_MIP_USE_PARTIALSOL_LEVEL}
\alias{LS_IPARAM_MIP_GENERAL_MODE}
\alias{LS_IPARAM_MIP_POLISH_NUM_BRANCH_NEXT}
\alias{LS_IPARAM_MIP_POLISH_MAX_BRANCH_COUNT}
\alias{LS_DPARAM_MIP_POLISH_ALPHA_TARGET}
\alias{LS_IPARAM_MIP_CONCURRENT_STRATEGY}
\alias{LS_DPARAM_MIP_BRANCH_TOP_VAL_DIFF_WEIGHT}
\alias{LS_IPARAM_MIP_BASCUTS_DONUM}
\alias{LS_IPARAM_MIP_PARA_SUB}
\alias{LS_DPARAM_MIP_PARA_RND_ITRLMT}
\alias{LS_DPARAM_MIP_PARA_INIT_NODE}
\alias{LS_IPARAM_MIP_PARA_ITR_MODE}
\alias{LS_IPARAM_MIP_PARA_FP}
\alias{LS_IPARAM_MIP_PARA_FP_MODE}
\alias{LS_IPARAM_MIP_HEU_DROP_OBJ}
\alias{LS_DPARAM_MIP_ABSCUTTOL}
\alias{LS_IPARAM_MIP_PERSPECTIVE_REFORM}
\alias{LS_IPARAM_MIP_TREEREORDERMODE}
\alias{LS_DPARAM_GOP_RELOPTTOL}
\alias{LS_DPARAM_GOP_FLTTOL}
\alias{LS_DPARAM_GOP_BOXTOL}
\alias{LS_DPARAM_GOP_WIDTOL}
\alias{LS_DPARAM_GOP_DELTATOL}
\alias{LS_DPARAM_GOP_BNDLIM}
\alias{LS_IPARAM_GOP_TIMLIM}
\alias{LS_IPARAM_GOP_OPTCHKMD}
\alias{LS_IPARAM_GOP_BRANCHMD}
\alias{LS_IPARAM_GOP_MAXWIDMD}
\alias{LS_IPARAM_GOP_PRELEVEL}
\alias{LS_IPARAM_GOP_POSTLEVEL}
\alias{LS_IPARAM_GOP_BBSRCHMD}
\alias{LS_IPARAM_GOP_DECOMPPTMD}
\alias{LS_IPARAM_GOP_ALGREFORMMD}
\alias{LS_IPARAM_GOP_RELBRNDMD}
\alias{LS_IPARAM_GOP_PRINTLEVEL}
\alias{LS_IPARAM_GOP_BNDLIM_MODE}
\alias{LS_IPARAM_GOP_BRANCH_LIMIT}
\alias{LS_IPARAM_GOP_CORELEVEL}
\alias{LS_IPARAM_GOP_OPT_MODE}
\alias{LS_IPARAM_GOP_HEU_MODE}
\alias{LS_IPARAM_GOP_SUBOUT_MODE}
\alias{LS_IPARAM_GOP_USE_NLPSOLVE}
\alias{LS_IPARAM_GOP_LSOLBRANLIM}
\alias{LS_IPARAM_GOP_LPSOPT}
\alias{LS_DPARAM_GOP_TIMLIM}
\alias{LS_DPARAM_GOP_BRANCH_LIMIT}
\alias{LS_IPARAM_GOP_QUADMD}
\alias{LS_IPARAM_GOP_LIM_MODE}
\alias{LS_DPARAM_GOP_ITRLIM}
\alias{LS_DPARAM_GOP_ITRLIM_SIM}
\alias{LS_DPARAM_GOP_ITRLIM_IPM}
\alias{LS_DPARAM_GOP_ITRLIM_NLP}
\alias{LS_DPARAM_GOP_ABSOPTTOL}
\alias{LS_DPARAM_GOP_PEROPTTOL}
\alias{LS_DPARAM_GOP_AOPTTIMLIM}
\alias{LS_IPARAM_GOP_LINEARZ}
\alias{LS_IPARAM_GOP_NUM_THREADS}
\alias{LS_IPARAM_GOP_MULTILINEAR}
\alias{LS_IPARAM_LIC_CONSTRAINTS}
\alias{LS_IPARAM_LIC_VARIABLES}
\alias{LS_IPARAM_LIC_INTEGERS}
\alias{LS_IPARAM_LIC_NONLINEARVARS}
\alias{LS_IPARAM_LIC_GOP_INTEGERS}
\alias{LS_IPARAM_LIC_GOP_NONLINEARVARS}
\alias{LS_IPARAM_LIC_DAYSTOEXP}
\alias{LS_IPARAM_LIC_DAYSTOTRIALEXP}
\alias{LS_IPARAM_LIC_NONLINEAR}
\alias{LS_IPARAM_LIC_EDUCATIONAL}
\alias{LS_IPARAM_LIC_RUNTIME}
\alias{LS_IPARAM_LIC_NUMUSERS}
\alias{LS_IPARAM_LIC_BARRIER}
\alias{LS_IPARAM_LIC_GLOBAL}
\alias{LS_IPARAM_LIC_PLATFORM}
\alias{LS_IPARAM_LIC_MIP}
\alias{LS_IPARAM_LIC_SP}
\alias{LS_IPARAM_LIC_CONIC}
\alias{LS_IPARAM_LIC_RESERVED1}
\alias{LS_IPARAM_IIS_ANALYZE_LEVEL}
\alias{LS_IPARAM_IUS_ANALYZE_LEVEL}
\alias{LS_IPARAM_IIS_TOPOPT}
\alias{LS_IPARAM_IIS_REOPT}
\alias{LS_IPARAM_IIS_USE_SFILTER}
\alias{LS_IPARAM_IIS_PRINT_LEVEL}
\alias{LS_IPARAM_IIS_INFEAS_NORM}
\alias{LS_IPARAM_IIS_ITER_LIMIT}
\alias{LS_DPARAM_IIS_ITER_LIMIT}
\alias{LS_IPARAM_IIS_TIME_LIMIT}
\alias{LS_IPARAM_IIS_METHOD}
\alias{LS_IPARAM_IIS_USE_EFILTER}
\alias{LS_IPARAM_IIS_USE_GOP}
\alias{LS_IPARAM_FMT_ISSQL}
\alias{LS_IPARAM_STOC_NSAMPLE_SPAR}
\alias{LS_IPARAM_STOC_NSAMPLE_STAGE}
\alias{LS_IPARAM_STOC_RG_SEED}
\alias{LS_IPARAM_STOC_METHOD}
\alias{LS_IPARAM_STOC_REOPT}
\alias{LS_IPARAM_STOC_TOPOPT}
\alias{LS_IPARAM_STOC_ITER_LIM}
\alias{LS_IPARAM_STOC_PRINT_LEVEL}
\alias{LS_IPARAM_STOC_DETEQ_TYPE}
\alias{LS_IPARAM_STOC_CALC_EVPI}
\alias{LS_IPARAM_STOC_SAMP_CONT_ONLY}
\alias{LS_IPARAM_STOC_BUCKET_SIZE}
\alias{LS_IPARAM_STOC_MAX_NUMSCENS}
\alias{LS_IPARAM_STOC_SHARE_BEGSTAGE}
\alias{LS_IPARAM_STOC_NODELP_PRELEVEL}
\alias{LS_DPARAM_STOC_TIME_LIM}
\alias{LS_DPARAM_STOC_RELOPTTOL}
\alias{LS_DPARAM_STOC_ABSOPTTOL}
\alias{LS_IPARAM_STOC_DEBUG_MASK}
\alias{LS_IPARAM_STOC_VARCONTROL_METHOD}
\alias{LS_IPARAM_STOC_CORRELATION_TYPE}
\alias{LS_IPARAM_STOC_WSBAS}
\alias{LS_IPARAM_STOC_ALD_OUTER_ITER_LIM}
\alias{LS_IPARAM_STOC_ALD_INNER_ITER_LIM}
\alias{LS_DPARAM_STOC_ALD_DUAL_FEASTOL}
\alias{LS_DPARAM_STOC_ALD_PRIMAL_FEASTOL}
\alias{LS_DPARAM_STOC_ALD_DUAL_STEPLEN}
\alias{LS_DPARAM_STOC_ALD_PRIMAL_STEPLEN}
\alias{LS_IPARAM_CORE_ORDER_BY_STAGE}
\alias{LS_SPARAM_STOC_FMT_NODE_NAME}
\alias{LS_SPARAM_STOC_FMT_SCENARIO_NAME}
\alias{LS_IPARAM_STOC_MAP_MPI2LP}
\alias{LS_IPARAM_STOC_AUTOAGGR}
\alias{LS_IPARAM_STOC_BENCHMARK_SCEN}
\alias{LS_DPARAM_STOC_INFBND}
\alias{LS_IPARAM_STOC_ADD_MPI}
\alias{LS_IPARAM_STOC_ELIM_FXVAR}
\alias{LS_DPARAM_STOC_SBD_OBJCUTVAL}
\alias{LS_IPARAM_STOC_SBD_OBJCUTFLAG}
\alias{LS_IPARAM_STOC_SBD_NUMCANDID}
\alias{LS_DPARAM_STOC_BIGM}
\alias{LS_IPARAM_STOC_NAMEDATA_LEVEL}
\alias{LS_IPARAM_STOC_SBD_MAXCUTS}
\alias{LS_IPARAM_STOC_DEQOPT}
\alias{LS_IPARAM_STOC_DS_SUBFORM}
\alias{LS_DPARAM_STOC_REL_PSTEPTOL}
\alias{LS_DPARAM_STOC_REL_DSTEPTOL}
\alias{LS_IPARAM_STOC_DETEQ_NBLOCKS}
\alias{LS_IPARAM_SAMP_NCM_METHOD}
\alias{LS_DPARAM_SAMP_NCM_CUTOBJ}
\alias{LS_IPARAM_SAMP_NCM_DSTORAGE}
\alias{LS_DPARAM_SAMP_CDSINC}
\alias{LS_IPARAM_SAMP_SCALE}
\alias{LS_IPARAM_SAMP_NCM_ITERLIM}
\alias{LS_DPARAM_SAMP_NCM_OPTTOL}
\alias{LS_IPARAM_SAMP_RG_BUFFER_SIZE}
\alias{LS_DPARAM_BNP_INFBND}
\alias{LS_IPARAM_BNP_LEVEL}
\alias{LS_IPARAM_BNP_PRINT_LEVEL}
\alias{LS_DPARAM_BNP_BOX_SIZE}
\alias{LS_DPARAM_BNP_SUB_ITRLMT}
\alias{LS_IPARAM_BNP_FIND_BLK}
\alias{LS_IPARAM_BNP_PRELEVEL}
\alias{LS_DPARAM_BNP_COL_LMT}
\alias{LS_DPARAM_BNP_TIMLIM}
\alias{LS_DPARAM_BNP_ITRLIM_SIM}
\alias{LS_DPARAM_BNP_ITRLIM_IPM}
\alias{LS_IPARAM_BNP_BRANCH_LIMIT}
\alias{LS_DPARAM_BNP_ITRLIM}
\alias{LS_DPARAM_GA_CXOVER_PROB}
\alias{LS_DPARAM_GA_XOVER_SPREAD}
\alias{LS_DPARAM_GA_IXOVER_PROB}
\alias{LS_DPARAM_GA_CMUTAT_PROB}
\alias{LS_DPARAM_GA_MUTAT_SPREAD}
\alias{LS_DPARAM_GA_IMUTAT_PROB}
\alias{LS_DPARAM_GA_TOL_ZERO}
\alias{LS_DPARAM_GA_TOL_PFEAS}
\alias{LS_DPARAM_GA_INF}
\alias{LS_DPARAM_GA_INFBND}
\alias{LS_DPARAM_GA_BLXA}
\alias{LS_DPARAM_GA_BLXB}
\alias{LS_IPARAM_GA_CXOVER_METHOD}
\alias{LS_IPARAM_GA_IXOVER_METHOD}
\alias{LS_IPARAM_GA_CMUTAT_METHOD}
\alias{LS_IPARAM_GA_IMUTAT_METHOD}
\alias{LS_IPARAM_GA_SEED}
\alias{LS_IPARAM_GA_NGEN}
\alias{LS_IPARAM_GA_POPSIZE}
\alias{LS_IPARAM_GA_FILEOUT}
\alias{LS_IPARAM_GA_PRINTLEVEL}
\alias{LS_IPARAM_GA_INJECT_OPT}
\alias{LS_IPARAM_GA_OBJDIR}
\alias{LS_DPARAM_GA_OBJSTOP}
\alias{LS_DPARAM_GA_MIGRATE_PROB}
\alias{LS_IPARAM_GA_SSPACE}
\alias{LS_IPARAM_VER_MAJOR}
\alias{LS_IPARAM_VER_MINOR}
\alias{LS_IPARAM_VER_BUILD}
\alias{LS_IPARAM_VER_REVISION}
\alias{LS_IPARAM_VER_NUMBER}
\alias{EP_NO_OP}
\alias{EP_PLUS}
\alias{EP_MINUS}
\alias{EP_MULTIPLY}
\alias{EP_DIVIDE}
\alias{EP_POWER}
\alias{EP_EQUAL}
\alias{EP_NOT_EQUAL}
\alias{EP_LTOREQ}
\alias{EP_GTOREQ}
\alias{EP_LTHAN}
\alias{EP_GTHAN}
\alias{EP_AND}
\alias{EP_OR}
\alias{EP_NOT}
\alias{EP_PERCENT}
\alias{EP_POSATE}
\alias{EP_NEGATE}
\alias{EP_ABS}
\alias{EP_SQRT}
\alias{EP_LOG}
\alias{EP_LN}
\alias{EP_PI}
\alias{EP_SIN}
\alias{EP_COS}
\alias{EP_TAN}
\alias{EP_ATAN2}
\alias{EP_ATAN}
\alias{EP_ASIN}
\alias{EP_ACOS}
\alias{EP_EXP}
\alias{EP_MOD}
\alias{EP_FALSE}
\alias{EP_TRUE}
\alias{EP_IF}
\alias{EP_PSN}
\alias{EP_PSL}
\alias{EP_LGM}
\alias{EP_SIGN}
\alias{EP_FLOOR}
\alias{EP_FPA}
\alias{EP_FPL}
\alias{EP_PEL}
\alias{EP_PEB}
\alias{EP_PPS}
\alias{EP_PPL}
\alias{EP_PTD}
\alias{EP_PCX}
\alias{EP_WRAP}
\alias{EP_PBNO}
\alias{EP_PFS}
\alias{EP_PFD}
\alias{EP_PHG}
\alias{EP_RAND}
\alias{EP_USER}
\alias{EP_SUM}
\alias{EP_AVG}
\alias{EP_MIN}
\alias{EP_MAX}
\alias{EP_NPV}
\alias{EP_VAND}
\alias{EP_VOR}
\alias{EP_PUSH_NUM}
\alias{EP_PUSH_VAR}
\alias{EP_NORMDENS}
\alias{EP_NORMINV}
\alias{EP_TRIAINV}
\alias{EP_EXPOINV}
\alias{EP_UNIFINV}
\alias{EP_MULTINV}
\alias{EP_USRCOD}
\alias{EP_SUMPROD}
\alias{EP_SUMIF}
\alias{EP_VLOOKUP}
\alias{EP_VPUSH_NUM}
\alias{EP_VPUSH_VAR}
\alias{EP_VMULT}
\alias{EP_SQR}
\alias{EP_SINH}
\alias{EP_COSH}
\alias{EP_TANH}
\alias{EP_ASINH}
\alias{EP_ACOSH}
\alias{EP_ATANH}
\alias{EP_LOGB}
\alias{EP_LOGX}
\alias{EP_LNX}
\alias{EP_TRUNC}
\alias{EP_NORMSINV}
\alias{EP_INT}
\alias{EP_PUSH_STR}
\alias{EP_VPUSH_STR}
\alias{EP_PUSH_SPAR}
\alias{EP_NORMPDF}
\alias{EP_NORMCDF}
\alias{EP_LSQ}
\alias{EP_LNPSNX}
\alias{EP_LNCPSN}
\alias{EP_XEXPNAX}
\alias{EP_XNEXPMX}
\alias{EP_PBT}
\alias{EP_PBTINV}
\alias{EP_PBNINV}
\alias{EP_PCC}
\alias{EP_PCCINV}
\alias{EP_PCXINV}
\alias{EP_EXPN}
\alias{EP_PFDINV}
\alias{EP_PGA}
\alias{EP_PGAINV}
\alias{EP_PGE}
\alias{EP_PGEINV}
\alias{EP_PGU}
\alias{EP_PGUINV}
\alias{EP_PHGINV}
\alias{EP_PLA}
\alias{EP_PLAINV}
\alias{EP_PLG}
\alias{EP_PLGINV}
\alias{EP_LGT}
\alias{EP_LGTINV}
\alias{EP_LGNM}
\alias{EP_LGNMINV}
\alias{EP_NGBN}
\alias{EP_NGBNINV}
\alias{EP_NRM}
\alias{EP_PPT}
\alias{EP_PPTINV}
\alias{EP_PPSINV}
\alias{EP_PTDINV}
\alias{EP_TRIAN}
\alias{EP_UNIFM}
\alias{EP_PWB}
\alias{EP_PWBINV}
\alias{EP_NRMINV}
\alias{EP_TRIANINV}
\alias{EP_EXPNINV}
\alias{EP_UNIFMINV}
\alias{EP_MLTNMINV}
\alias{EP_BTDENS}
\alias{EP_BNDENS}
\alias{EP_CCDENS}
\alias{EP_CXDENS}
\alias{EP_EXPDENS}
\alias{EP_FDENS}
\alias{EP_GADENS}
\alias{EP_GEDENS}
\alias{EP_GUDENS}
\alias{EP_HGDENS}
\alias{EP_LADENS}
\alias{EP_LGDENS}
\alias{EP_LGTDENS}
\alias{EP_LGNMDENS}
\alias{EP_NGBNDENS}
\alias{EP_NRMDENS}
\alias{EP_PTDENS}
\alias{EP_PSDENS}
\alias{EP_TDENS}
\alias{EP_TRIADENS}
\alias{EP_UNIFDENS}
\alias{EP_WBDENS}
\alias{EP_RADIANS}
\alias{EP_DEGREES}
\alias{EP_ROUND}
\alias{EP_ROUNDUP}
\alias{EP_ROUNDDOWN}
\alias{EP_ERF}
\alias{EP_PBN}
\alias{EP_PBB}
\alias{EP_PBBINV}
\alias{EP_BBDENS}
\alias{EP_PSS}
\alias{EP_SSDENS}
\alias{EP_SSINV}
\alias{EP_POSD}
\alias{LS_IINFO_NUM_NONZ_OBJ}
\alias{LS_IINFO_NUM_SEMICONT}
\alias{LS_IINFO_NUM_SETS}
\alias{LS_IINFO_NUM_SETS_NNZ}
\alias{LS_IINFO_NUM_QCP_CONS}
\alias{LS_IINFO_NUM_CONT_CONS}
\alias{LS_IINFO_NUM_INT_CONS}
\alias{LS_IINFO_NUM_BIN_CONS}
\alias{LS_IINFO_NUM_QCP_VARS}
\alias{LS_IINFO_NUM_CONS}
\alias{LS_IINFO_NUM_VARS}
\alias{LS_IINFO_NUM_NONZ}
\alias{LS_IINFO_NUM_BIN}
\alias{LS_IINFO_NUM_INT}
\alias{LS_IINFO_NUM_CONT}
\alias{LS_IINFO_NUM_QC_NONZ}
\alias{LS_IINFO_NUM_NLP_NONZ}
\alias{LS_IINFO_NUM_NLPOBJ_NONZ}
\alias{LS_IINFO_NUM_RDCONS}
\alias{LS_IINFO_NUM_RDVARS}
\alias{LS_IINFO_NUM_RDNONZ}
\alias{LS_IINFO_NUM_RDINT}
\alias{LS_IINFO_LEN_VARNAMES}
\alias{LS_IINFO_LEN_CONNAMES}
\alias{LS_IINFO_NUM_NLP_CONS}
\alias{LS_IINFO_NUM_NLP_VARS}
\alias{LS_IINFO_NUM_SUF_ROWS}
\alias{LS_IINFO_NUM_IIS_ROWS}
\alias{LS_IINFO_NUM_SUF_BNDS}
\alias{LS_IINFO_NUM_IIS_BNDS}
\alias{LS_IINFO_NUM_SUF_COLS}
\alias{LS_IINFO_NUM_IUS_COLS}
\alias{LS_IINFO_NUM_CONES}
\alias{LS_IINFO_NUM_CONE_NONZ}
\alias{LS_IINFO_LEN_CONENAMES}
\alias{LS_DINFO_INST_VAL_MIN_COEF}
\alias{LS_IINFO_INST_VARNDX_MIN_COEF}
\alias{LS_IINFO_INST_CONNDX_MIN_COEF}
\alias{LS_DINFO_INST_VAL_MAX_COEF}
\alias{LS_IINFO_INST_VARNDX_MAX_COEF}
\alias{LS_IINFO_INST_CONNDX_MAX_COEF}
\alias{LS_IINFO_NUM_VARS_CARD}
\alias{LS_IINFO_NUM_VARS_SOS1}
\alias{LS_IINFO_NUM_VARS_SOS2}
\alias{LS_IINFO_NUM_VARS_SOS3}
\alias{LS_IINFO_NUM_VARS_SCONT}
\alias{LS_IINFO_NUM_CONS_L}
\alias{LS_IINFO_NUM_CONS_E}
\alias{LS_IINFO_NUM_CONS_G}
\alias{LS_IINFO_NUM_CONS_R}
\alias{LS_IINFO_NUM_CONS_N}
\alias{LS_IINFO_NUM_VARS_LB}
\alias{LS_IINFO_NUM_VARS_UB}
\alias{LS_IINFO_NUM_VARS_LUB}
\alias{LS_IINFO_NUM_VARS_FR}
\alias{LS_IINFO_NUM_VARS_FX}
\alias{LS_IINFO_NUM_INST_CODES}
\alias{LS_IINFO_NUM_INST_REAL_NUM}
\alias{LS_IINFO_NUM_SPARS}
\alias{LS_IINFO_NUM_PROCS}
\alias{LS_IINFO_NUM_POSDS}
\alias{LS_IINFO_METHOD}
\alias{LS_DINFO_POBJ}
\alias{LS_DINFO_DOBJ}
\alias{LS_DINFO_PINFEAS}
\alias{LS_DINFO_DINFEAS}
\alias{LS_IINFO_MODEL_STATUS}
\alias{LS_IINFO_PRIMAL_STATUS}
\alias{LS_IINFO_DUAL_STATUS}
\alias{LS_IINFO_BASIC_STATUS}
\alias{LS_IINFO_BAR_ITER}
\alias{LS_IINFO_SIM_ITER}
\alias{LS_IINFO_NLP_ITER}
\alias{LS_IINFO_ELAPSED_TIME}
\alias{LS_DINFO_MSW_POBJ}
\alias{LS_IINFO_MSW_PASS}
\alias{LS_IINFO_MSW_NSOL}
\alias{LS_IINFO_IPM_STATUS}
\alias{LS_DINFO_IPM_POBJ}
\alias{LS_DINFO_IPM_DOBJ}
\alias{LS_DINFO_IPM_PINFEAS}
\alias{LS_DINFO_IPM_DINFEAS}
\alias{LS_IINFO_NLP_CALL_FUN}
\alias{LS_IINFO_NLP_CALL_DEV}
\alias{LS_IINFO_NLP_CALL_HES}
\alias{LS_IINFO_CONCURRENT_OPTIMIZER}
\alias{LS_IINFO_LEN_STAGENAMES}
\alias{LS_DINFO_BAR_ITER}
\alias{LS_DINFO_SIM_ITER}
\alias{LS_DINFO_NLP_ITER}
\alias{LS_IINFO_BAR_THREADS}
\alias{LS_IINFO_NLP_THREADS}
\alias{LS_IINFO_SIM_THREADS}
\alias{LS_DINFO_NLP_THRIMBL}
\alias{LS_SINFO_NLP_THREAD_LOAD}
\alias{LS_SINFO_BAR_THREAD_LOAD}
\alias{LS_SINFO_SIM_THREAD_LOAD}
\alias{LS_SINFO_ARCH}
\alias{LS_IINFO_ARCH_ID}
\alias{LS_IINFO_MSW_BESTRUNIDX}
\alias{LS_DINFO_ACONDEST}
\alias{LS_DINFO_BCONDEST}
\alias{LS_IINFO_LPTOOL}
\alias{LS_IINFO_MIP_STATUS}
\alias{LS_DINFO_MIP_OBJ}
\alias{LS_DINFO_MIP_BESTBOUND}
\alias{LS_IINFO_MIP_SIM_ITER}
\alias{LS_IINFO_MIP_BAR_ITER}
\alias{LS_IINFO_MIP_NLP_ITER}
\alias{LS_IINFO_MIP_BRANCHCOUNT}
\alias{LS_IINFO_MIP_NEWIPSOL}
\alias{LS_IINFO_MIP_LPCOUNT}
\alias{LS_IINFO_MIP_ACTIVENODES}
\alias{LS_IINFO_MIP_LTYPE}
\alias{LS_IINFO_MIP_AOPTTIMETOSTOP}
\alias{LS_IINFO_MIP_NUM_TOTAL_CUTS}
\alias{LS_IINFO_MIP_GUB_COVER_CUTS}
\alias{LS_IINFO_MIP_FLOW_COVER_CUTS}
\alias{LS_IINFO_MIP_LIFT_CUTS}
\alias{LS_IINFO_MIP_PLAN_LOC_CUTS}
\alias{LS_IINFO_MIP_DISAGG_CUTS}
\alias{LS_IINFO_MIP_KNAPSUR_COVER_CUTS}
\alias{LS_IINFO_MIP_LATTICE_CUTS}
\alias{LS_IINFO_MIP_GOMORY_CUTS}
\alias{LS_IINFO_MIP_COEF_REDC_CUTS}
\alias{LS_IINFO_MIP_GCD_CUTS}
\alias{LS_IINFO_MIP_OBJ_CUT}
\alias{LS_IINFO_MIP_BASIS_CUTS}
\alias{LS_IINFO_MIP_CARDGUB_CUTS}
\alias{LS_IINFO_MIP_CLIQUE_CUTS}
\alias{LS_IINFO_MIP_CONTRA_CUTS}
\alias{LS_IINFO_MIP_GUB_CONS}
\alias{LS_IINFO_MIP_GLB_CONS}
\alias{LS_IINFO_MIP_PLANTLOC_CONS}
\alias{LS_IINFO_MIP_DISAGG_CONS}
\alias{LS_IINFO_MIP_SB_CONS}
\alias{LS_IINFO_MIP_IKNAP_CONS}
\alias{LS_IINFO_MIP_KNAP_CONS}
\alias{LS_IINFO_MIP_NLP_CONS}
\alias{LS_IINFO_MIP_CONT_CONS}
\alias{LS_DINFO_MIP_TOT_TIME}
\alias{LS_DINFO_MIP_OPT_TIME}
\alias{LS_DINFO_MIP_HEU_TIME}
\alias{LS_IINFO_MIP_SOLSTATUS_LAST_BRANCH}
\alias{LS_DINFO_MIP_SOLOBJVAL_LAST_BRANCH}
\alias{LS_IINFO_MIP_HEU_LEVEL}
\alias{LS_DINFO_MIP_PFEAS}
\alias{LS_DINFO_MIP_INTPFEAS}
\alias{LS_IINFO_MIP_WHERE_IN_CODE}
\alias{LS_IINFO_MIP_FP_ITER}
\alias{LS_DINFO_MIP_FP_SUMFEAS}
\alias{LS_DINFO_MIP_RELMIPGAP}
\alias{LS_DINFO_MIP_ROOT_OPT_TIME}
\alias{LS_DINFO_MIP_ROOT_PRE_TIME}
\alias{LS_IINFO_MIP_ROOT_METHOD}
\alias{LS_DINFO_MIP_SIM_ITER}
\alias{LS_DINFO_MIP_BAR_ITER}
\alias{LS_DINFO_MIP_NLP_ITER}
\alias{LS_IINFO_MIP_TOP_RELAX_IS_NON_CONVEX}
\alias{LS_DINFO_MIP_FP_TIME}
\alias{LS_IINFO_MIP_THREADS}
\alias{LS_SINFO_MIP_THREAD_LOAD}
\alias{LS_DINFO_MIP_ABSGAP}
\alias{LS_DINFO_MIP_RELGAP}
\alias{LS_IINFO_MIP_SOFTKNAP_CUTS}
\alias{LS_DINFO_GOP_OBJ}
\alias{LS_IINFO_GOP_SIM_ITER}
\alias{LS_IINFO_GOP_BAR_ITER}
\alias{LS_IINFO_GOP_NLP_ITER}
\alias{LS_DINFO_GOP_BESTBOUND}
\alias{LS_IINFO_GOP_STATUS}
\alias{LS_IINFO_GOP_LPCOUNT}
\alias{LS_IINFO_GOP_NLPCOUNT}
\alias{LS_IINFO_GOP_MIPCOUNT}
\alias{LS_IINFO_GOP_NEWSOL}
\alias{LS_IINFO_GOP_BOX}
\alias{LS_IINFO_GOP_BBITER}
\alias{LS_IINFO_GOP_SUBITER}
\alias{LS_IINFO_GOP_MIPBRANCH}
\alias{LS_IINFO_GOP_ACTIVEBOXES}
\alias{LS_IINFO_GOP_TOT_TIME}
\alias{LS_IINFO_GOP_MAXDEPTH}
\alias{LS_DINFO_GOP_PFEAS}
\alias{LS_DINFO_GOP_INTPFEAS}
\alias{LS_DINFO_GOP_SIM_ITER}
\alias{LS_DINFO_GOP_BAR_ITER}
\alias{LS_DINFO_GOP_NLP_ITER}
\alias{LS_DINFO_GOP_LPCOUNT}
\alias{LS_DINFO_GOP_NLPCOUNT}
\alias{LS_DINFO_GOP_MIPCOUNT}
\alias{LS_DINFO_GOP_BBITER}
\alias{LS_DINFO_GOP_SUBITER}
\alias{LS_DINFO_GOP_MIPBRANCH}
\alias{LS_DINFO_GOP_FIRST_TIME}
\alias{LS_DINFO_GOP_BEST_TIME}
\alias{LS_DINFO_GOP_TOT_TIME}
\alias{LS_IINFO_GOP_THREADS}
\alias{LS_SINFO_GOP_THREAD_LOAD}
\alias{LS_DINFO_GOP_ABSGAP}
\alias{LS_DINFO_GOP_RELGAP}
\alias{LS_DINFO_SUB_OBJ}
\alias{LS_DINFO_SUB_PINF}
\alias{LS_DINFO_CUR_OBJ}
\alias{LS_IINFO_CUR_ITER}
\alias{LS_DINFO_CUR_BEST_BOUND}
\alias{LS_IINFO_CUR_STATUS}
\alias{LS_IINFO_CUR_LP_COUNT}
\alias{LS_IINFO_CUR_BRANCH_COUNT}
\alias{LS_IINFO_CUR_ACTIVE_COUNT}
\alias{LS_IINFO_CUR_NLP_COUNT}
\alias{LS_IINFO_CUR_MIP_COUNT}
\alias{LS_IINFO_CUR_CUT_COUNT}
\alias{LS_DINFO_CUR_ITER}
\alias{LS_DINFO_GEN_PERCENT}
\alias{LS_IINFO_GEN_NONZ_TTL}
\alias{LS_IINFO_GEN_NONZ_NL}
\alias{LS_IINFO_GEN_ROW_NL}
\alias{LS_IINFO_GEN_VAR_NL}
\alias{LS_IINFO_IIS_BAR_ITER}
\alias{LS_IINFO_IIS_SIM_ITER}
\alias{LS_IINFO_IIS_NLP_ITER}
\alias{LS_DINFO_IIS_BAR_ITER}
\alias{LS_DINFO_IIS_SIM_ITER}
\alias{LS_DINFO_IIS_NLP_ITER}
\alias{LS_IINFO_IIS_TOT_TIME}
\alias{LS_IINFO_IIS_ACT_NODE}
\alias{LS_IINFO_IIS_LPCOUNT}
\alias{LS_IINFO_IIS_NLPCOUNT}
\alias{LS_IINFO_IIS_MIPCOUNT}
\alias{LS_IINFO_IIS_THREADS}
\alias{LS_SINFO_IIS_THREAD_LOAD}
\alias{LS_IINFO_IUS_BAR_ITER}
\alias{LS_IINFO_IUS_SIM_ITER}
\alias{LS_IINFO_IUS_NLP_ITER}
\alias{LS_DINFO_IUS_BAR_ITER}
\alias{LS_DINFO_IUS_SIM_ITER}
\alias{LS_DINFO_IUS_NLP_ITER}
\alias{LS_IINFO_IUS_TOT_TIME}
\alias{LS_IINFO_IUS_ACT_NODE}
\alias{LS_IINFO_IUS_LPCOUNT}
\alias{LS_IINFO_IUS_NLPCOUNT}
\alias{LS_IINFO_IUS_MIPCOUNT}
\alias{LS_IINFO_IUS_THREADS}
\alias{LS_SINFO_IUS_THREAD_LOAD}
\alias{LS_IINFO_PRE_NUM_RED}
\alias{LS_IINFO_PRE_TYPE_RED}
\alias{LS_IINFO_PRE_NUM_RDCONS}
\alias{LS_IINFO_PRE_NUM_RDVARS}
\alias{LS_IINFO_PRE_NUM_RDNONZ}
\alias{LS_IINFO_PRE_NUM_RDINT}
\alias{LS_IINFO_ERR_OPTIM}
\alias{LS_DINFO_PROFILE_BASE}
\alias{LS_DINFO_PROFILE_IIS_FIND_NEC_ROWS}
\alias{LS_DINFO_PROFILE_IIS_FIND_NEC_COLS}
\alias{LS_DINFO_PROFILE_IIS_FIND_SUF_ROWS}
\alias{LS_DINFO_PROFILE_IIS_FIND_SUF_COLS}
\alias{LS_DINFO_PROFILE_MIP_ROOT_LP}
\alias{LS_DINFO_PROFILE_MIP_TOTAL_LP}
\alias{LS_DINFO_PROFILE_MIP_LP_SIM_PRIMAL}
\alias{LS_DINFO_PROFILE_MIP_LP_SIM_DUAL}
\alias{LS_DINFO_PROFILE_MIP_LP_SIM_BARRIER}
\alias{LS_DINFO_PROFILE_MIP_PRE_PROCESS}
\alias{LS_DINFO_PROFILE_MIP_FEA_PUMP}
\alias{LS_DINFO_PROFILE_MIP_TOP_HEURISTIC}
\alias{LS_DINFO_PROFILE_MIP_BNB_HEURISTIC}
\alias{LS_DINFO_PROFILE_MIP_BNB_MAIN_LOOP}
\alias{LS_DINFO_PROFILE_MIP_BNB_SUB_LOOP}
\alias{LS_DINFO_PROFILE_MIP_BNB_BEFORE_BEST}
\alias{LS_DINFO_PROFILE_MIP_BNB_AFTER_BEST}
\alias{LS_DINFO_PROFILE_MIP_TOP_CUT}
\alias{LS_DINFO_PROFILE_MIP_BNB_CUT}
\alias{LS_DINFO_PROFILE_MIP_LP_NON_BNB_LOOP}
\alias{LS_DINFO_PROFILE_MIP_LP_BNB_LOOP_MAIN}
\alias{LS_DINFO_PROFILE_MIP_LP_BNB_LOOP_SUB}
\alias{LS_DINFO_PROFILE_MIP_NODE_PRESOLVE}
\alias{LS_DINFO_PROFILE_MIP_BNB_BRANCHING}
\alias{LS_DINFO_PROFILE_MIP_BNB_BRANCHING_MAIN}
\alias{LS_DINFO_PROFILE_MIP_BNB_BRANCHING_SUB}
\alias{LS_DINFO_PROFILE_GOP_SUB_LP_SOLVER}
\alias{LS_DINFO_PROFILE_GOP_SUB_NLP_SOLVER}
\alias{LS_DINFO_PROFILE_GOP_SUB_MIP_SOLVER}
\alias{LS_DINFO_PROFILE_GOP_SUB_GOP_SOLVER}
\alias{LS_DINFO_PROFILE_GOP_CONS_PROP_LP}
\alias{LS_DINFO_PROFILE_GOP_CONS_PROP_NLP}
\alias{LS_DINFO_PROFILE_GOP_VAR_MIN_MAX}
\alias{LS_SINFO_MODEL_FILENAME}
\alias{LS_SINFO_MODEL_SOURCE}
\alias{LS_IINFO_MODEL_TYPE}
\alias{LS_SINFO_CORE_FILENAME}
\alias{LS_SINFO_STOC_FILENAME}
\alias{LS_SINFO_TIME_FILENAME}
\alias{LS_IINFO_ASSIGNED_MODEL_TYPE}
\alias{LS_DINFO_STOC_EVOBJ}
\alias{LS_DINFO_STOC_EVPI}
\alias{LS_DINFO_STOC_PINFEAS}
\alias{LS_DINFO_STOC_DINFEAS}
\alias{LS_DINFO_STOC_RELOPT_GAP}
\alias{LS_DINFO_STOC_ABSOPT_GAP}
\alias{LS_IINFO_STOC_SIM_ITER}
\alias{LS_IINFO_STOC_BAR_ITER}
\alias{LS_IINFO_STOC_NLP_ITER}
\alias{LS_IINFO_NUM_STOCPAR_RHS}
\alias{LS_IINFO_NUM_STOCPAR_OBJ}
\alias{LS_IINFO_NUM_STOCPAR_LB}
\alias{LS_IINFO_NUM_STOCPAR_UB}
\alias{LS_IINFO_NUM_STOCPAR_INSTR_OBJS}
\alias{LS_IINFO_NUM_STOCPAR_INSTR_CONS}
\alias{LS_IINFO_NUM_STOCPAR_AIJ}
\alias{LS_DINFO_STOC_TOTAL_TIME}
\alias{LS_IINFO_STOC_STATUS}
\alias{LS_IINFO_STOC_STAGE_BY_NODE}
\alias{LS_IINFO_STOC_NUM_SCENARIOS}
\alias{LS_DINFO_STOC_NUM_SCENARIOS}
\alias{LS_IINFO_STOC_NUM_STAGES}
\alias{LS_IINFO_STOC_NUM_NODES}
\alias{LS_DINFO_STOC_NUM_NODES}
\alias{LS_IINFO_STOC_NUM_NODES_STAGE}
\alias{LS_DINFO_STOC_NUM_NODES_STAGE}
\alias{LS_IINFO_STOC_NUM_NODE_MODELS}
\alias{LS_IINFO_STOC_NUM_COLS_BEFORE_NODE}
\alias{LS_IINFO_STOC_NUM_ROWS_BEFORE_NODE}
\alias{LS_IINFO_STOC_NUM_COLS_DETEQI}
\alias{LS_DINFO_STOC_NUM_COLS_DETEQI}
\alias{LS_IINFO_STOC_NUM_ROWS_DETEQI}
\alias{LS_DINFO_STOC_NUM_ROWS_DETEQI}
\alias{LS_IINFO_STOC_NUM_COLS_DETEQE}
\alias{LS_DINFO_STOC_NUM_COLS_DETEQE}
\alias{LS_IINFO_STOC_NUM_ROWS_DETEQE}
\alias{LS_DINFO_STOC_NUM_ROWS_DETEQE}
\alias{LS_IINFO_STOC_NUM_COLS_NAC}
\alias{LS_IINFO_STOC_NUM_ROWS_NAC}
\alias{LS_IINFO_STOC_NUM_COLS_CORE}
\alias{LS_IINFO_STOC_NUM_ROWS_CORE}
\alias{LS_IINFO_STOC_NUM_COLS_STAGE}
\alias{LS_IINFO_STOC_NUM_ROWS_STAGE}
\alias{LS_IINFO_STOC_NUM_NBF_CUTS}
\alias{LS_IINFO_STOC_NUM_NBO_CUTS}
\alias{LS_IINFO_DIST_TYPE}
\alias{LS_IINFO_SAMP_SIZE}
\alias{LS_DINFO_SAMP_MEAN}
\alias{LS_DINFO_SAMP_STD}
\alias{LS_DINFO_SAMP_SKEWNESS}
\alias{LS_DINFO_SAMP_KURTOSIS}
\alias{LS_IINFO_STOC_NUM_QCP_CONS_DETEQE}
\alias{LS_IINFO_STOC_NUM_CONT_CONS_DETEQE}
\alias{LS_IINFO_STOC_NUM_INT_CONS_DETEQE}
\alias{LS_IINFO_STOC_NUM_BIN_CONS_DETEQE}
\alias{LS_IINFO_STOC_NUM_QCP_VARS_DETEQE}
\alias{LS_IINFO_STOC_NUM_NONZ_DETEQE}
\alias{LS_IINFO_STOC_NUM_BIN_DETEQE}
\alias{LS_IINFO_STOC_NUM_INT_DETEQE}
\alias{LS_IINFO_STOC_NUM_CONT_DETEQE}
\alias{LS_IINFO_STOC_NUM_QC_NONZ_DETEQE}
\alias{LS_IINFO_STOC_NUM_NLP_NONZ_DETEQE}
\alias{LS_IINFO_STOC_NUM_NLPOBJ_NONZ_DETEQE}
\alias{LS_IINFO_STOC_NUM_QCP_CONS_DETEQI}
\alias{LS_IINFO_STOC_NUM_CONT_CONS_DETEQI}
\alias{LS_IINFO_STOC_NUM_INT_CONS_DETEQI}
\alias{LS_IINFO_STOC_NUM_BIN_CONS_DETEQI}
\alias{LS_IINFO_STOC_NUM_QCP_VARS_DETEQI}
\alias{LS_IINFO_STOC_NUM_NONZ_DETEQI}
\alias{LS_IINFO_STOC_NUM_BIN_DETEQI}
\alias{LS_IINFO_STOC_NUM_INT_DETEQI}
\alias{LS_IINFO_STOC_NUM_CONT_DETEQI}
\alias{LS_IINFO_STOC_NUM_QC_NONZ_DETEQI}
\alias{LS_IINFO_STOC_NUM_NLP_NONZ_DETEQI}
\alias{LS_IINFO_STOC_NUM_NLPOBJ_NONZ_DETEQI}
\alias{LS_IINFO_STOC_NUM_EVENTS_BLOCK}
\alias{LS_IINFO_STOC_NUM_EVENTS_DISCRETE}
\alias{LS_IINFO_STOC_NUM_EVENTS_PARAMETRIC}
\alias{LS_IINFO_STOC_NUM_EXPLICIT_SCENARIOS}
\alias{LS_IINFO_STOC_PARENT_NODE}
\alias{LS_IINFO_STOC_ELDEST_CHILD_NODE}
\alias{LS_IINFO_STOC_NUM_CHILD_NODES}
\alias{LS_IINFO_NUM_STOCPAR_INSTR}
\alias{LS_IINFO_INFORUNB_SCEN_IDX}
\alias{LS_DINFO_STOC_EVMU}
\alias{LS_DINFO_STOC_EVWS}
\alias{LS_DINFO_STOC_EVAVR}
\alias{LS_IINFO_DIST_NARG}
\alias{LS_IINFO_SAMP_VARCONTROL_METHOD}
\alias{LS_IINFO_STOC_NUM_NLP_VARS_DETEQE}
\alias{LS_IINFO_STOC_NUM_NLP_CONS_DETEQE}
\alias{LS_DINFO_STOC_EVOBJ_LB}
\alias{LS_DINFO_STOC_EVOBJ_UB}
\alias{LS_DINFO_STOC_AVROBJ}
\alias{LS_DINFO_SAMP_MEDIAN}
\alias{LS_DINFO_DIST_MEDIAN}
\alias{LS_IINFO_STOC_NUM_CC}
\alias{LS_IINFO_STOC_NUM_ROWS_CC}
\alias{LS_IINFO_STOC_ISCBACK}
\alias{LS_IINFO_STOC_LP_COUNT}
\alias{LS_IINFO_STOC_NLP_COUNT}
\alias{LS_IINFO_STOC_MIP_COUNT}
\alias{LS_DINFO_STOC_OPT_TIME}
\alias{LS_DINFO_SAMP_CORRDIFF_ST}
\alias{LS_DINFO_SAMP_CORRDIFF_CT}
\alias{LS_DINFO_SAMP_CORRDIFF_SC}
\alias{LS_IINFO_STOC_NUM_EQROWS_CC}
\alias{LS_IINFO_STOC_NUM_ROWS}
\alias{LS_IINFO_STOC_NUM_CC_VIOLATED}
\alias{LS_IINFO_STOC_NUM_COLS_DETEQC}
\alias{LS_IINFO_STOC_NUM_ROWS_DETEQC}
\alias{LS_IINFO_STOC_NUM_QCP_CONS_DETEQC}
\alias{LS_IINFO_STOC_NUM_CONT_CONS_DETEQC}
\alias{LS_IINFO_STOC_NUM_INT_CONS_DETEQC}
\alias{LS_IINFO_STOC_NUM_BIN_CONS_DETEQC}
\alias{LS_IINFO_STOC_NUM_QCP_VARS_DETEQC}
\alias{LS_IINFO_STOC_NUM_NONZ_DETEQC}
\alias{LS_IINFO_STOC_NUM_BIN_DETEQC}
\alias{LS_IINFO_STOC_NUM_INT_DETEQC}
\alias{LS_IINFO_STOC_NUM_CONT_DETEQC}
\alias{LS_IINFO_STOC_NUM_QC_NONZ_DETEQC}
\alias{LS_IINFO_STOC_NUM_NLP_NONZ_DETEQC}
\alias{LS_IINFO_STOC_NUM_NLPOBJ_NONZ_DETEQC}
\alias{LS_IINFO_STOC_NUM_NLP_CONS_DETEQC}
\alias{LS_IINFO_STOC_NUM_NLP_VARS_DETEQC}
\alias{LS_IINFO_STOC_NUM_NONZ_OBJ_DETEQC}
\alias{LS_IINFO_STOC_NUM_NONZ_OBJ_DETEQE}
\alias{LS_DINFO_STOC_CC_PLEVEL}
\alias{LS_IINFO_STOC_THREADS}
\alias{LS_DINFO_STOC_THRIMBL}
\alias{LS_IINFO_STOC_NUM_EQROWS}
\alias{LS_SINFO_STOC_THREAD_LOAD}
\alias{LS_IINFO_STOC_NUM_BUCKETS}
\alias{LS_IINFO_BNP_SIM_ITER}
\alias{LS_IINFO_BNP_LPCOUNT}
\alias{LS_IINFO_BNP_NUMCOL}
\alias{LS_DINFO_BNP_BESTBOUND}
\alias{LS_DINFO_BNP_BESTOBJ}
\alias{LSERR_NO_ERROR}
\alias{LSERR_OUT_OF_MEMORY}
\alias{LSERR_CANNOT_OPEN_FILE}
\alias{LSERR_BAD_MPS_FILE}
\alias{LSERR_BAD_CONSTRAINT_TYPE}
\alias{LSERR_BAD_MODEL}
\alias{LSERR_BAD_SOLVER_TYPE}
\alias{LSERR_BAD_OBJECTIVE_SENSE}
\alias{LSERR_BAD_MPI_FILE}
\alias{LSERR_INFO_NOT_AVAILABLE}
\alias{LSERR_ILLEGAL_NULL_POINTER}
\alias{LSERR_UNABLE_TO_SET_PARAM}
\alias{LSERR_INDEX_OUT_OF_RANGE}
\alias{LSERR_ERRMSG_FILE_NOT_FOUND}
\alias{LSERR_VARIABLE_NOT_FOUND}
\alias{LSERR_INTERNAL_ERROR}
\alias{LSERR_ITER_LIMIT}
\alias{LSERR_TIME_LIMIT}
\alias{LSERR_NOT_CONVEX}
\alias{LSERR_NUMERIC_INSTABILITY}
\alias{LSERR_STEP_TOO_SMALL}
\alias{LSERR_USER_INTERRUPT}
\alias{LSERR_PARAMETER_OUT_OF_RANGE}
\alias{LSERR_ERROR_IN_INPUT}
\alias{LSERR_TOO_SMALL_LICENSE}
\alias{LSERR_NO_VALID_LICENSE}
\alias{LSERR_NO_METHOD_LICENSE}
\alias{LSERR_NOT_SUPPORTED}
\alias{LSERR_MODEL_ALREADY_LOADED}
\alias{LSERR_MODEL_NOT_LOADED}
\alias{LSERR_INDEX_DUPLICATE}
\alias{LSERR_INSTRUCT_NOT_LOADED}
\alias{LSERR_OLD_LICENSE}
\alias{LSERR_NO_LICENSE_FILE}
\alias{LSERR_BAD_LICENSE_FILE}
\alias{LSERR_MIP_BRANCH_LIMIT}
\alias{LSERR_GOP_FUNC_NOT_SUPPORTED}
\alias{LSERR_GOP_BRANCH_LIMIT}
\alias{LSERR_BAD_DECOMPOSITION_TYPE}
\alias{LSERR_BAD_VARIABLE_TYPE}
\alias{LSERR_BASIS_BOUND_MISMATCH}
\alias{LSERR_BASIS_COL_STATUS}
\alias{LSERR_BASIS_INVALID}
\alias{LSERR_BASIS_ROW_STATUS}
\alias{LSERR_BLOCK_OF_BLOCK}
\alias{LSERR_BOUND_OUT_OF_RANGE}
\alias{LSERR_COL_BEGIN_INDEX}
\alias{LSERR_COL_INDEX_OUT_OF_RANGE}
\alias{LSERR_COL_NONZCOUNT}
\alias{LSERR_INVALID_ERRORCODE}
\alias{LSERR_ROW_INDEX_OUT_OF_RANGE}
\alias{LSERR_TOTAL_NONZCOUNT}
\alias{LSERR_MODEL_NOT_LINEAR}
\alias{LSERR_CHECKSUM}
\alias{LSERR_USER_FUNCTION_NOT_FOUND}
\alias{LSERR_TRUNCATED_NAME_DATA}
\alias{LSERR_ILLEGAL_STRING_OPERATION}
\alias{LSERR_STRING_ALREADY_LOADED}
\alias{LSERR_STRING_NOT_LOADED}
\alias{LSERR_STRING_LENGTH_LIMIT}
\alias{LSERR_DATA_TERM_EXIST}
\alias{LSERR_NOT_SORTED_ORDER}
\alias{LSERR_INST_MISS_ELEMENTS}
\alias{LSERR_INST_TOO_SHORT}
\alias{LSERR_INST_INVALID_BOUND}
\alias{LSERR_INST_SYNTAX_ERROR}
\alias{LSERR_COL_TOKEN_NOT_FOUND}
\alias{LSERR_ROW_TOKEN_NOT_FOUND}
\alias{LSERR_NAME_TOKEN_NOT_FOUND}
\alias{LSERR_NOT_LSQ_MODEL}
\alias{LSERR_INCOMPATBLE_DECOMPOSITION}
\alias{LSERR_NO_MULTITHREAD_SUPPORT}
\alias{LSERR_INVALID_PARAMID}
\alias{LSERR_INVALID_NTHREADS}
\alias{LSERR_COL_LIMIT}
\alias{LSERR_QCDATA_NOT_LOADED}
\alias{LSERR_NO_QCDATA_IN_ROW}
\alias{LSERR_CLOCK_SETBACK}
\alias{LSERR_LDL_FACTORIZATION}
\alias{LSERR_LDL_EMPTY_COL}
\alias{LSERR_LDL_BAD_MATRIX_DATA}
\alias{LSERR_LDL_INVALID_DIM}
\alias{LSERR_LDL_EMPTY_MATRIX}
\alias{LSERR_LDL_MATRIX_NOTSYM}
\alias{LSERR_LDL_ZERO_DIAG}
\alias{LSERR_LDL_INVALID_PERM}
\alias{LSERR_LDL_DUPELEM}
\alias{LSERR_LDL_RANK}
\alias{LSERR_BAD_SMPS_CORE_FILE}
\alias{LSERR_BAD_SMPS_TIME_FILE}
\alias{LSERR_BAD_SMPS_STOC_FILE}
\alias{LSERR_BAD_SMPI_CORE_FILE}
\alias{LSERR_BAD_SMPI_STOC_FILE}
\alias{LSERR_CANNOT_OPEN_CORE_FILE}
\alias{LSERR_CANNOT_OPEN_TIME_FILE}
\alias{LSERR_CANNOT_OPEN_STOC_FILE}
\alias{LSERR_STOC_MODEL_NOT_LOADED}
\alias{LSERR_STOC_SPAR_NOT_FOUND}
\alias{LSERR_TIME_SPAR_NOT_FOUND}
\alias{LSERR_SCEN_INDEX_OUT_OF_SEQUENCE}
\alias{LSERR_STOC_MODEL_ALREADY_PARSED}
\alias{LSERR_STOC_INVALID_SCENARIO_CDF}
\alias{LSERR_CORE_SPAR_NOT_FOUND}
\alias{LSERR_CORE_SPAR_COUNT_MISMATCH}
\alias{LSERR_CORE_INVALID_SPAR_INDEX}
\alias{LSERR_TIME_SPAR_NOT_EXPECTED}
\alias{LSERR_TIME_SPAR_COUNT_MISMATCH}
\alias{LSERR_CORE_SPAR_VALUE_NOT_FOUND}
\alias{LSERR_INFO_UNAVAILABLE}
\alias{LSERR_STOC_MISSING_BNDNAME}
\alias{LSERR_STOC_MISSING_OBJNAME}
\alias{LSERR_STOC_MISSING_RHSNAME}
\alias{LSERR_STOC_MISSING_RNGNAME}
\alias{LSERR_MISSING_TOKEN_NAME}
\alias{LSERR_MISSING_TOKEN_ROOT}
\alias{LSERR_STOC_NODE_UNBOUNDED}
\alias{LSERR_STOC_NODE_INFEASIBLE}
\alias{LSERR_STOC_TOO_MANY_SCENARIOS}
\alias{LSERR_STOC_BAD_PRECISION}
\alias{LSERR_CORE_BAD_AGGREGATION}
\alias{LSERR_STOC_NULL_EVENT_TREE}
\alias{LSERR_CORE_BAD_STAGE_INDEX}
\alias{LSERR_STOC_BAD_ALGORITHM}
\alias{LSERR_CORE_BAD_NUMSTAGES}
\alias{LSERR_TIME_BAD_TEMPORAL_ORDER}
\alias{LSERR_TIME_BAD_NUMSTAGES}
\alias{LSERR_CORE_TIME_MISMATCH}
\alias{LSERR_STOC_INVALID_CDF}
\alias{LSERR_BAD_DISTRIBUTION_TYPE}
\alias{LSERR_DIST_SCALE_OUT_OF_RANGE}
\alias{LSERR_DIST_SHAPE_OUT_OF_RANGE}
\alias{LSERR_DIST_INVALID_PROBABILITY}
\alias{LSERR_DIST_NO_DERIVATIVE}
\alias{LSERR_DIST_INVALID_SD}
\alias{LSERR_DIST_INVALID_X}
\alias{LSERR_DIST_INVALID_PARAMS}
\alias{LSERR_DIST_ROOTER_ITERLIM}
\alias{LSERR_ARRAY_OUT_OF_BOUNDS}
\alias{LSERR_DIST_NO_PDF_LIMIT}
\alias{LSERR_RG_NOT_SET}
\alias{LSERR_DIST_TRUNCATED}
\alias{LSERR_STOC_MISSING_PARAM_TOKEN}
\alias{LSERR_DIST_INVALID_NUMPARAM}
\alias{LSERR_CORE_NOT_IN_TEMPORAL_ORDER}
\alias{LSERR_STOC_INVALID_SAMPLE_SIZE}
\alias{LSERR_STOC_NOT_DISCRETE}
\alias{LSERR_STOC_SCENARIO_LIMIT}
\alias{LSERR_DIST_BAD_CORRELATION_TYPE}
\alias{LSERR_TIME_NUMSTAGES_NOT_SET}
\alias{LSERR_STOC_SAMPLE_ALREADY_LOADED}
\alias{LSERR_STOC_EVENTS_NOT_LOADED}
\alias{LSERR_STOC_TREE_ALREADY_INIT}
\alias{LSERR_RG_SEED_NOT_SET}
\alias{LSERR_STOC_OUT_OF_SAMPLE_POINTS}
\alias{LSERR_STOC_SCENARIO_SAMPLING_NOT_SUPPORTED}
\alias{LSERR_STOC_SAMPLE_NOT_GENERATED}
\alias{LSERR_STOC_SAMPLE_ALREADY_GENERATED}
\alias{LSERR_STOC_SAMPLE_SIZE_TOO_SMALL}
\alias{LSERR_RG_ALREADY_SET}
\alias{LSERR_STOC_BLOCK_SAMPLING_NOT_SUPPORTED}
\alias{LSERR_EMPTY_SPAR_STAGE}
\alias{LSERR_EMPTY_ROW_STAGE}
\alias{LSERR_EMPTY_COL_STAGE}
\alias{LSERR_STOC_CONFLICTING_SAMP_SIZES}
\alias{LSERR_STOC_EMPTY_SCENARIO_DATA}
\alias{LSERR_STOC_CORRELATION_NOT_INDUCED}
\alias{LSERR_STOC_PDF_TABLE_NOT_LOADED}
\alias{LSERR_STOC_NO_CONTINUOUS_SPAR_FOUND}
\alias{LSERR_STOC_ROW_ALREADY_IN_CC}
\alias{LSERR_STOC_CC_NOT_LOADED}
\alias{LSERR_STOC_CUT_LIMIT}
\alias{LSERR_STOC_GA_NOT_INIT}
\alias{LSERR_STOC_ROWS_NOT_LOADED_IN_CC}
\alias{LSERR_SAMP_ALREADY_SOURCE}
\alias{LSERR_SAMP_USERFUNC_NOT_SET}
\alias{LSERR_SAMP_INVALID_CALL}
\alias{LSERR_STOC_MAP_MULTI_SPAR}
\alias{LSERR_STOC_MAP_SAME_SPAR}
\alias{LSERR_STOC_SPAR_NOT_EXPECTED_OBJ}
\alias{LSERR_DIST_PARAM_NOT_SET}
\alias{LSERR_SPRINT_MISSING_TAG_ROWS}
\alias{LSERR_SPRINT_MISSING_TAG_COLS}
\alias{LSERR_SPRINT_MISSING_TAG_RHS}
\alias{LSERR_SPRINT_MISSING_TAG_ENDATA}
\alias{LSERR_SPRINT_MISSING_VALUE_ROW}
\alias{LSERR_SPRINT_EXTRA_VALUE_ROW}
\alias{LSERR_SPRINT_MISSING_VALUE_COL}
\alias{LSERR_SPRINT_EXTRA_VALUE_COL}
\alias{LSERR_SPRINT_MISSING_VALUE_RHS}
\alias{LSERR_SPRINT_EXTRA_VALUE_RHS}
\alias{LSERR_SPRINT_MISSING_VALUE_BOUND}
\alias{LSERR_SPRINT_EXTRA_VALUE_BOUND}
\alias{LSERR_SPRINT_INTEGER_VARS_IN_MPS}
\alias{LSERR_SPRINT_BINARY_VARS_IN_MPS}
\alias{LSERR_SPRINT_SEMI_CONT_VARS_IN_MPS}
\alias{LSERR_SPRINT_UNKNOWN_TAG_BOUNDS}
\alias{LSERR_SPRINT_MULTIPLE_OBJ_ROWS}
\alias{LSERR_SPRINT_COULD_NOT_SOLVE_SUBPROBLEM}
\alias{LSERR_COULD_NOT_WRITE_TO_FILE}
\alias{LSERR_COULD_NOT_READ_FROM_FILE}
\alias{LSERR_READING_PAST_EOF}
\alias{LSERR_LAST_ERROR}
\alias{LS_METHOD_FREE}
\alias{LS_METHOD_PSIMPLEX}
\alias{LS_METHOD_DSIMPLEX}
\alias{LS_METHOD_BARRIER}
\alias{LS_METHOD_NLP}
\alias{LS_METHOD_MIP}
\alias{LS_METHOD_MULTIS}
\alias{LS_METHOD_GOP}
\alias{LS_METHOD_IIS}
\alias{LS_METHOD_IUS}
\alias{LS_METHOD_SBD}
\alias{LS_METHOD_GA}
\alias{LS_STRATEGY_USER}
\alias{LS_STRATEGY_PRIMIP}
\alias{LS_STRATEGY_NODEMIP}
\alias{LS_STRATEGY_HEUMIP}
\alias{LS_NMETHOD_FREE}
\alias{LS_NMETHOD_LSQ}
\alias{LS_NMETHOD_QP}
\alias{LS_NMETHOD_CONOPT}
\alias{LS_NMETHOD_SLP}
\alias{LS_NMETHOD_MSW_GRG}
\alias{LS_PROB_SOLVE_FREE}
\alias{LS_PROB_SOLVE_PRIMAL}
\alias{LS_PROB_SOLVE_DUAL}
\alias{LS_BAR_METHOD_FREE}
\alias{LS_BAR_METHOD_INTPNT}
\alias{LS_BAR_METHOD_CONIC}
\alias{LS_BAR_METHOD_QCONE}
\alias{LSSOL_BASIC_PRIMAL}
\alias{LSSOL_BASIC_DUAL}
\alias{LSSOL_BASIC_SLACK}
\alias{LSSOL_BASIC_REDCOST}
\alias{LSSOL_INTERIOR_PRIMAL}
\alias{LSSOL_INTERIOR_DUAL}
\alias{LSSOL_INTERIOR_SLACK }
\alias{LSSOL_INTERIOR_REDCOST}
\alias{LS_LP}
\alias{LS_QP}
\alias{LS_SOCP}
\alias{LS_SDP}
\alias{LS_NLP}
\alias{LS_MILP}
\alias{LS_MIQP}
\alias{LS_MISOCP}
\alias{LS_MISDP}
\alias{LS_MINLP}
\alias{LS_CONVEX_QP}
\alias{LS_CONVEX_NLP}
\alias{LS_CONVEX_MIQP}
\alias{LS_CONVEX_MINLP}
\alias{LS_UNDETERMINED}
\alias{LS_LINK_BLOCKS_FREE}
\alias{LS_LINK_BLOCKS_SELF}
\alias{LS_LINK_BLOCKS_NONE}
\alias{LS_LINK_BLOCKS_COLS}
\alias{LS_LINK_BLOCKS_ROWS}
\alias{LS_LINK_BLOCKS_BOTH }
\alias{LS_LINK_BLOCKS_MATRIX}
\alias{LS_MPS_USE_MAX_NOTE}
\alias{LS_MPS_USE_MAX_CARD}
\alias{LS_MPS_USE_MAX_FLIP}
\alias{LS_DERIV_FREE}
\alias{LS_DERIV_FORWARD_DIFFERENCE}
\alias{LS_DERIV_BACKWARD_DIFFERENCE}
\alias{LS_DERIV_CENTER_DIFFERENCE}
\alias{LS_MIP_SET_CARD}
\alias{LS_MIP_SET_SOS1}
\alias{LS_MIP_SET_SOS2}
\alias{LS_MIP_SET_SOS3}
\alias{LS_QTERM_NONE}
\alias{LS_QTERM_INDEF}
\alias{LS_QTERM_POSDEF}
\alias{LS_QTERM_NEGDEF}
\alias{LS_QTERM_POS_SEMIDEF}
\alias{LS_QTERM_NEG_SEMIDEF}
\alias{LS_MIP_MODE_NO_TIME_EVENTS}
\alias{LS_MIP_MODE_FAST_FEASIBILITY}
\alias{LS_MIP_MODE_FAST_OPTIMALITY}
\alias{LS_MIP_MODE_NO_BRANCH_CUTS}
\alias{LS_MIP_GUB_COVER_CUTS}
\alias{LS_MIP_FLOW_COVER_CUTS}
\alias{LS_MIP_LIFT_CUTS}
\alias{LS_MIP_PLAN_LOC_CUTS}
\alias{LS_MIP_DISAGG_CUTS}
\alias{LS_MIP_KNAPSUR_COVER_CUTS}
\alias{LS_MIP_LATTICE_CUTS}
\alias{LS_MIP_GOMORY_CUTS}
\alias{LS_MIP_COEF_REDC_CUTS}
\alias{LS_MIP_GCD_CUTS}
\alias{LS_MIP_OBJ_CUT}
\alias{LS_MIP_BASIS_CUTS}
\alias{LS_MIP_CARDGUB_CUTS}
\alias{LS_MIP_DISJUN_CUTS}
\alias{LS_MIP_PREP_SPRE}
\alias{LS_MIP_PREP_PROB}
\alias{LS_MIP_PREP_COEF}
\alias{LS_MIP_PREP_ELIM}
\alias{LS_MIP_PREP_DUAL}
\alias{LS_MIP_PREP_DBACK}
\alias{LS_MIP_PREP_BINROWS}
\alias{LS_MIP_PREP_AGGROWS}
\alias{LS_MIP_PREP_COEF_LIFTING}
\alias{LS_MIP_PREP_MAXPASS}
\alias{LS_SOLVER_PREP_SPRE}
\alias{LS_SOLVER_PREP_PFOR}
\alias{LS_SOLVER_PREP_DFOR}
\alias{LS_SOLVER_PREP_ELIM}
\alias{LS_SOLVER_PREP_DCOL}
\alias{LS_SOLVER_PREP_DROW}
\alias{LS_SOLVER_PREP_MAXPASS}
\alias{LS_NECESSARY_ROWS}
\alias{LS_NECESSARY_COLS}
\alias{LS_SUFFICIENT_ROWS}
\alias{LS_SUFFICIENT_COLS}
\alias{LS_IIS_INTS}
\alias{LS_IISRANK_LTF}
\alias{LS_IISRANK_DECOMP}
\alias{LS_IISRANK_NNZ}
\alias{LS_IISLIMIT_MIS}
\alias{LS_IIS_NORM_FREE}
\alias{LS_IIS_NORM_ONE}
\alias{LS_IIS_NORM_INFINITY}
\alias{LS_IIS_DEFAULT}
\alias{LS_IIS_DEL_FILTER}
\alias{LS_IIS_ADD_FILTER}
\alias{LS_IIS_GBS_FILTER}
\alias{LS_IIS_DFBS_FILTER}
\alias{LS_IIS_FSC_FILTER}
\alias{LS_IIS_ELS_FILTER}
\alias{LS_MIP_IN_PRESOLVE}
\alias{LS_MIP_IN_FP_MODE}
\alias{LS_MIP_IN_HEU_MODE}
\alias{LS_MIP_IN_ENUM}
\alias{LS_MIP_IN_CUT_ADD_TOP}
\alias{LS_MIP_IN_CUT_ADD_TREE}
\alias{LS_MIP_IN_BANDB}
\alias{LS_JCOL_INST}
\alias{LS_JCOL_RUB}
\alias{LS_JCOL_RLB}
\alias{LS_JCOL_RHS}
\alias{LS_IROW_OBJ}
\alias{LS_IROW_VUB}
\alias{LS_IROW_VLB}
\alias{LS_IROW_VFX}
\alias{LS_IMAT_AIJ}
\alias{LSDIST_TYPE_BINOMIAL}
\alias{LSDIST_TYPE_DISCRETE}
\alias{LSDIST_TYPE_DISCRETE_BLOCK}
\alias{LSDIST_TYPE_NEGATIVE_BINOMIAL}
\alias{LSDIST_TYPE_GEOMETRIC}
\alias{LSDIST_TYPE_POISSON}
\alias{LSDIST_TYPE_LOGARITHMIC}
\alias{LSDIST_TYPE_HYPER_GEOMETRIC}
\alias{LSDIST_TYPE_LINTRAN_BLOCK}
\alias{LSDIST_TYPE_SUB_BLOCK}
\alias{LSDIST_TYPE_SUB}
\alias{LSDIST_TYPE_USER}
\alias{LSDIST_TYPE_BETA}
\alias{LSDIST_TYPE_CAUCHY}
\alias{LSDIST_TYPE_CHI_SQUARE}
\alias{LSDIST_TYPE_EXPONENTIAL}
\alias{LSDIST_TYPE_F_DISTRIBUTION}
\alias{LSDIST_TYPE_GAMMA}
\alias{LSDIST_TYPE_GUMBEL}
\alias{LSDIST_TYPE_LAPLACE}
\alias{LSDIST_TYPE_LOGNORMAL}
\alias{LSDIST_TYPE_LOGISTIC}
\alias{LSDIST_TYPE_NORMAL}
\alias{LSDIST_TYPE_PARETO}
\alias{LSDIST_TYPE_STABLE_PARETIAN}
\alias{LSDIST_TYPE_STUDENTS_T}
\alias{LSDIST_TYPE_TRIANGULAR}
\alias{LSDIST_TYPE_UNIFORM}
\alias{LSDIST_TYPE_WEIBULL}
\alias{LSDIST_TYPE_WILCOXON}
\alias{LSDIST_TYPE_BETABINOMIAL}
\alias{LSDIST_TYPE_SYMMETRICSTABLE}
\alias{LS_REPLACE}
\alias{LS_ADD}
\alias{LS_SUB}
\alias{LS_MULTIPLY}
\alias{LS_DIVIDE}
\alias{LS_SCEN_ROOT}
\alias{LS_SCEN_AVRG}
\alias{LS_SCEN_MEDIAN}
\alias{LS_SCEN_USER}
\alias{LS_SCEN_NONE}
\alias{LS_WSBAS_FREE}
\alias{LS_WSBAS_NONE}
\alias{LS_WSBAS_AVRG}
\alias{LS_WSBAS_LAST}
\alias{LS_METHOD_STOC_FREE}
\alias{LS_METHOD_STOC_DETEQ}
\alias{LS_METHOD_STOC_NBD}
\alias{LS_METHOD_STOC_ALD}
\alias{LS_METHOD_STOC_HS}
\alias{LS_DETEQ_FREE}
\alias{LS_DETEQ_IMPLICIT}
\alias{LS_DETEQ_EXPLICIT}
\alias{LS_DETEQ_CHANCE}
\alias{LS_USER}
\alias{LS_PDF}
\alias{LS_CDF}
\alias{LS_CDFINV}
\alias{LS_PDFDIFF}
\alias{LS_CORR_TARGET}
\alias{LS_CORR_LINEAR}
\alias{LS_CORR_PEARSON}
\alias{LS_CORR_KENDALL}
\alias{LS_CORR_SPEARMAN}
\alias{LS_MONTECARLO}
\alias{LS_LATINSQUARE}
\alias{LS_ANTITHETIC}
\alias{LS_RANDGEN_FREE}
\alias{LS_RANDGEN_SYSTEM}
\alias{LS_RANDGEN_LINDO1}
\alias{LS_RANDGEN_LINDO2}
\alias{LS_RANDGEN_LIN1}
\alias{LS_RANDGEN_MULT1}
\alias{LS_RANDGEN_MULT2}
\alias{LS_RANDGEN_MERSENNE}
\alias{LS_NCM_STD}
\alias{LS_NCM_GA}
\alias{LS_NCM_ALTP}
\alias{LS_NCM_L2NORM_CONE}
\alias{LS_NCM_L2NORM_NLP}
\alias{LS_PTR_ENV}
\alias{LS_PTR_MODEL}
\alias{LS_PTR_SAMPLE}
\alias{LS_PTR_RG}
\alias{LS_MTMODE_FREE}
\alias{LS_MTMODE_EXPLCT}
\alias{LS_MTMODE_PPCC}
\alias{LS_MTMODE_PP}
\alias{LS_MTMODE_CCPP}
\alias{LS_MTMODE_CC}
\alias{LS_SPRINT_OUTPUT_FILE_FREE}
\alias{LS_SPRINT_OUTPUT_FILE_BIN}
\alias{LS_SPRINT_OUTPUT_FILE_TXT}
\alias{LS_MSW_MODE_TRUNCATE_FREE}
\alias{LS_MSW_MODE_SCALE_REFSET}
\alias{LS_MSW_MODE_EXPAND_RADIUS}
\alias{LS_MSW_MODE_SKEWED_SAMPLE}
\alias{LS_MSW_MODE_BEST_LOCAL_BND}
\alias{LS_MSW_MODE_BEST_GLOBAL_BND}
\alias{LS_MSW_MODE_SAMPLE_FREEVARS}
\alias{LS_MSW_MODE_PRECOLLECT}
\alias{LS_MSW_MODE_POWER_SOLVE}
\alias{LS_GA_CROSS_SBX}
\alias{LS_GA_CROSS_BLXA}
\alias{LS_GA_CROSS_BLXAB}
\alias{LS_GA_CROSS_HEU}
\alias{LS_GA_CROSS_ONEPOINT}
\alias{LS_GA_CROSS_TWOPOINT}

\title{
   LINDO API Parameters.
}

\description{
 For more information, please refer to LINDO API User Manual.
}

\references{
 LINDO SYSTEMS home page at www.lindo.com
}
